backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
python-jose[cryptography]
pytest
httpx
python-multipart
bcrypt==3.2.0
pydantic-settings
hyperliquid-python-sdk
websockets
